test: verify state sync via Prometheus metrics in e2e tests (#5786)
## Add Prometheus metrics verification to state sync e2e tests

Ref:
https://github.com/celestiaorg/celestia-app/issues/5447#issuecomment-3304391957

State sync e2e tests were only verifying that nodes reached the target
height and stopped catching up, but weren't confirming that actual state
sync (vs block sync) was used. This could lead to false positives where
block sync was used instead of state sync. This PR proposes to add check
Prometheus `apply_snapshot_chunk` metrics to confirm snapshot chunks
were actually applied.
